DEVPROPTYPE is ULONG which maps to C#'s uint https://docs.microsoft.com/en-us/windows/win32/winprog/windows-data-types#ulong

+++ src/HidLibrary/HidDevices.cs
@@ -137,7 +137,7 @@ namespace HidLibrary
 
             if (Environment.OSVersion.Version.Major > 5)
             {
-                ulong propertyType = 0;
+                uint propertyType = 0;
                 var requiredSize = 0;
 
                 var _continue = NativeMethods.SetupDiGetDeviceProperty(deviceInfoSet,
